Extracted service tests to folder
