+++ src/app/models/AbstractModel.spec.ts
@@ -2,12 +2,15 @@ import { HttpClientTestingModule } from "@angular/common/http/testing";
 import { Injector } from "@angular/core";
 import { async, fakeAsync, TestBed, tick } from "@angular/core/testing";
 import { ApiErrorDetails } from "@baw-api/api.interceptor.service";
-import { shouldNotFail, shouldNotSucceed } from "@baw-api/baw-api.service.spec";
 import { MockModel as ChildModel } from "@baw-api/mock/baseApiMock.service";
 import {
   MOCK,
   MockStandardApiService,
 } from "@baw-api/mock/standardApiMock.service";
+import {
+  shouldNotFail,
+  shouldNotSucceed,
+} from "@baw-api/tests/baw-api.service.spec";
 import { Id, Ids } from "@interfaces/apiInterfaces";
 import { DateTime, Duration } from "luxon";
 import { BehaviorSubject, Observable, Subject } from "rxjs";
