+++ src/app/services/baw-api/tests/account.service.spec.ts
@@ -1,8 +1,10 @@
 import { HttpClientTestingModule } from "@angular/common/http/testing";
 import { TestBed } from "@angular/core/testing";
 import { RouterTestingModule } from "@angular/router/testing";
+import { AccountService } from "@baw-api/account.service";
+import { BawApiService } from "@baw-api/baw-api.service";
+import { MockBawApiService } from "@baw-api/mock/baseApiMock.service";
 import { User } from "@models/User";
-import { testAppInitializer } from "src/app/test/helpers/testbed";
 import {
   validateApiCreate,
   validateApiDestroy,
@@ -10,10 +12,8 @@ import {
   validateApiList,
   validateApiShow,
   validateApiUpdate,
-} from "../../test/helpers/api-common";
-import { AccountService } from "./account.service";
-import { BawApiService } from "./baw-api.service";
-import { MockBawApiService } from "./mock/baseApiMock.service";
+} from "src/app/test/helpers/api-common";
+import { testAppInitializer } from "src/app/test/helpers/testbed";
 
 describe("AccountService", () => {
   beforeEach(function () {
