+++ src/app/services/baw-api/tests/analysis-jobs.service.spec.ts
@@ -2,8 +2,6 @@ import { HttpClientTestingModule } from "@angular/common/http/testing";
 import { TestBed } from "@angular/core/testing";
 import { RouterTestingModule } from "@angular/router/testing";
 import { AnalysisJob } from "@models/AnalysisJob";
-import { testAppInitializer } from "src/app/test.helper";
-import { AnalysisJobsService } from "./analysis-jobs.service";
 import {
   validateApiCreate,
   validateApiDestroy,
@@ -11,9 +9,11 @@ import {
   validateApiList,
   validateApiShow,
   validateApiUpdate,
-} from "./api-common.helper";
-import { BawApiService } from "./baw-api.service";
-import { MockBawApiService } from "./mock/baseApiMock.service";
+} from "src/app/test/helpers/api-common";
+import { testAppInitializer } from "src/app/test/helpers/testbed";
+import { AnalysisJobsService } from "../analysis-jobs.service";
+import { BawApiService } from "../baw-api.service";
+import { MockBawApiService } from "../mock/baseApiMock.service";
 
 describe("AnalysisJobsService", function () {
   beforeEach(function () {
