+++ src/app/services/baw-api/tests/api.interceptor.service.spec.ts
@@ -9,11 +9,14 @@ import {
   HttpTestingController,
 } from "@angular/common/http/testing";
 import { TestBed } from "@angular/core/testing";
+import {
+  ApiErrorDetails,
+  BawApiInterceptor,
+} from "@baw-api/api.interceptor.service";
+import { SecurityService } from "@baw-api/security.service";
 import { SessionUser } from "@models/User";
+import { AppConfigService } from "@services/app-config/app-config.service";
 import { testBawServices } from "src/app/test/helpers/testbed";
-import { AppConfigService } from "../app-config/app-config.service";
-import { ApiErrorDetails, BawApiInterceptor } from "./api.interceptor.service";
-import { SecurityService } from "./security.service";
 
 describe("BawApiInterceptor", () => {
   let api: SecurityService;
