+++ src/app/services/baw-api/tests/responses-shallow.service.spec.ts
@@ -2,7 +2,6 @@ import { HttpClientTestingModule } from "@angular/common/http/testing";
 import { TestBed } from "@angular/core/testing";
 import { RouterTestingModule } from "@angular/router/testing";
 import { Response } from "@models/Response";
-import { testAppInitializer } from "src/app/test.helper";
 import {
   validateApiCreate,
   validateApiDestroy,
@@ -10,11 +9,11 @@ import {
   validateApiList,
   validateApiShow,
   validateApiUpdate,
-} from "./api-common.helper";
-import { BawApiService } from "./baw-api.service";
-import { MockBawApiService } from "./mock/baseApiMock.service";
-import { ShallowQuestionsService } from "./questions.service";
-import { ShallowResponsesService } from "./responses.service";
+} from "src/app/test/helpers/api-common";
+import { testAppInitializer } from "src/app/test/helpers/testbed";
+import { BawApiService } from "../baw-api.service";
+import { MockBawApiService } from "../mock/baseApiMock.service";
+import { ShallowResponsesService } from "../responses.service";
 
 describe("ShallowResponsesService", function () {
   beforeEach(function () {
