+++ src/app/services/baw-api/tests/scripts.service.spec.ts
@@ -1,18 +1,18 @@
 import { HttpClientTestingModule } from "@angular/common/http/testing";
 import { async, TestBed } from "@angular/core/testing";
 import { RouterTestingModule } from "@angular/router/testing";
+import { BawApiService } from "@baw-api/baw-api.service";
+import { MockBawApiService } from "@baw-api/mock/baseApiMock.service";
+import { ScriptsService } from "@baw-api/scripts.service";
 import { Script } from "@models/Script";
-import { testAppInitializer } from "src/app/test/helpers/testbed";
 import {
   validateApiCreate,
   validateApiDestroy,
   validateApiFilter,
   validateApiList,
   validateApiShow,
-} from "../../test/helpers/api-common";
-import { BawApiService } from "./baw-api.service";
-import { MockBawApiService } from "./mock/baseApiMock.service";
-import { ScriptsService } from "./scripts.service";
+} from "src/app/test/helpers/api-common";
+import { testAppInitializer } from "src/app/test/helpers/testbed";
 
 describe("ScriptsService", function () {
   beforeEach(async(function () {
