+++ src/app/services/baw-api/tests/user.service.spec.ts
@@ -2,11 +2,11 @@ import { HttpClientTestingModule } from "@angular/common/http/testing";
 import { TestBed } from "@angular/core/testing";
 import { RouterTestingModule } from "@angular/router/testing";
 import { User } from "@models/User";
+import { validateApiShow } from "src/app/test/helpers/api-common";
 import { testAppInitializer } from "src/app/test/helpers/testbed";
-import { validateApiShow } from "../../test/helpers/api-common";
-import { BawApiService } from "./baw-api.service";
-import { MockBawApiService } from "./mock/baseApiMock.service";
-import { UserService } from "./user.service";
+import { BawApiService } from "../baw-api.service";
+import { MockBawApiService } from "../mock/baseApiMock.service";
+import { UserService } from "../user.service";
 
 describe("UserService", function () {
   beforeEach(function () {
